refactor(publikator): update pages folder to allow for routing with next.js routing
